Commit initial connection plans
Adding these to the commit history as I plan on changing them. It's
not clear what the best solution is, so keeping this snapshot of the
planned interface around sounded like a good idea.

Signed-off-by: David Ross <David.Ross@wdc.com>

Connection attribute reading spec 2

This still isn't the thing I'm actually going to implement, but saving
it seems reasonable.

Signed-off-by: David Ross <David.Ross@wdc.com>

More in-progress work towards connection/attribute support.

Signed-off-by: David Ross <David.Ross@wdc.com>